# src/apps/backend/src/OpenApiGenerator/src/Org.OpenAPITools/Controllers/AuthApi.cs
/*
 * ToDo Application API
 *
 * ToDoアプリのAPI設計
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{
    /// <summary>
    ///
    /// </summary>
    [ApiController]
    public abstract class AuthApiController : ControllerBase
    {
        /// <summary>
        /// ログイン
        /// </summary>
        /// <remarks>メールアドレスとパスワードを使用してユーザーを認証し、アクセストークンを返す</remarks>
        /// <param name="userLoginRequestDto"></param>
        /// <response code="200">ログイン成功</response>
        /// <response code="400">400 Bad Request</response>
        /// <response code="401">401 Unauthorized</response>
        /// <response code="404">404 Not Found</response>
        [HttpPost]
        [Route("/api/v1/auth/login")]
        [Consumes("application/json")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(UserLoginResponseDto))]
        [ProducesResponseType(statusCode: 400, type: typeof(ErrorResponseDto))]
        [ProducesResponseType(statusCode: 401, type: typeof(ErrorResponseDto))]
        [ProducesResponseType(statusCode: 404, type: typeof(ErrorResponseDto))]
        public abstract Task<IActionResult> Login([FromBody] UserLoginRequestDto userLoginRequestDto);
    }
}