# src/apps/backend/src/OpenApiGenerator/src/Org.OpenAPITools/Controllers/TasksApi.cs
/*
 * ToDo Application API
 *
 * ToDoアプリのAPI設計
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{
    /// <summary>
    ///
    /// </summary>
    [ApiController]
    public abstract class TasksApiController : ControllerBase
    {
        /// <summary>
        /// タスク削除
        /// </summary>
        /// <param name="taskId">UUID</param>
        /// <response code="200">200 OK</response>
        /// <response code="400">400 Bad Request</response>
        /// <response code="404">404 Not Found</response>
        [HttpDelete]
        [Route("/api/v1/tasks/{taskId}")]
        [Authorize]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 400, type: typeof(ErrorResponseDto))]
        [ProducesResponseType(statusCode: 404, type: typeof(ErrorResponseDto))]
        public abstract Task<IActionResult> DeleteTask([FromRoute(Name = "taskId")][Required] Guid taskId);

        /// <summary>
        /// タスク1件取得
        /// </summary>
        /// <param name="taskId">UUID</param>
        /// <response code="200">200 OK</response>
        /// <response code="400">400 Bad Request</response>
        /// <response code="404">404 Not Found</response>
        [HttpGet]
        [Route("/api/v1/tasks/{taskId}")]
        [Authorize]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(TaskGetResponseDto))]
        [ProducesResponseType(statusCode: 400, type: typeof(ErrorResponseDto))]
        [ProducesResponseType(statusCode: 404, type: typeof(ErrorResponseDto))]
        public abstract Task<IActionResult> GetTask([FromRoute(Name = "taskId")][Required] Guid taskId);

        /// <summary>
        /// タスク一覧取得
        /// </summary>
        /// <remarks>0件のときは空の配列</remarks>
        /// <response code="200">200 OK</response>
        [HttpGet]
        [Route("/api/v1/tasks")]
        [Authorize]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(List<TaskGetResponseDto>))]
        public abstract Task<IActionResult> GetTasks();

        /// <summary>
        /// タスク新規作成
        /// </summary>
        /// <param name="taskPostRequestDto"></param>
        /// <response code="201">201 Created</response>
        [HttpPost]
        [Route("/api/v1/tasks")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        public abstract Task<IActionResult> PostTask([FromBody] TaskPostRequestDto taskPostRequestDto);

        /// <summary>
        /// タスク更新
        /// </summary>
        /// <param name="taskId">UUID</param>
        /// <param name="taskPutRequestDto"></param>
        /// <response code="200">200 OK</response>
        /// <response code="400">400 Bad Request</response>
        /// <response code="404">404 Not Found</response>
        [HttpPut]
        [Route("/api/v1/tasks/{taskId}")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 400, type: typeof(ErrorResponseDto))]
        [ProducesResponseType(statusCode: 404, type: typeof(ErrorResponseDto))]
        public abstract Task<IActionResult> PutTask([FromRoute(Name = "taskId")][Required] Guid taskId, [FromBody] TaskPutRequestDto taskPutRequestDto);
    }
}